Split ScoreScriptUtilsTests into two separate classes.
